separate dense slam and enable IO of extrinsics
